Title: Instantiate and use StoryTellerFacade in StoryTeller.js
Description:
In this commit, the `StoryTeller.js` file is introduced in the `src/public/` directory. It imports the necessary classes required for the functionality of the Gutenberg Storyteller Block, such as Fetcher, SliderFactory, UIHandler, DomHelper, and the StoryTellerFacade.

The following steps are taken in the `StoryTeller.js` file:

1. Instantiate the required classes: Fetcher, SliderFactory, UIHandler, and DomHelper.
2. Instantiate the StoryTellerFacade class by passing the instances of the required classes to its constructor.
3. Attach click event listeners to the grid items by using the 'storyteller-grid-item' class.

When a grid item is clicked, the `initCategorySwiper` method of the StoryTellerFacade instance is called with the index of the clicked grid item as an argument. This initializes the category swiper for the clicked grid item.

This commit sets up the main entry point of the Gutenberg Storyteller Block by utilizing the StoryTellerFacade class and connecting it to the user interface.